% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{accuracy}
\alias{accuracy}
\title{Calculating accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{A vector that records prediction result}

\item{truth}{A vector that records ground truth label}
}
\value{
A number type result
}
\description{
Calculating accuracy
}
