% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculating adjusted R-squared statistics}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A vector that records the prediction results}

\item{truth}{A vector that records the ground truth label}
}
\value{
A number type result
}
\description{
Calculating adjusted R-squared statistics
}
