% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Finding the classes of each variable}
\usage{
classes(x)
}
\arguments{
\item{x}{Tibble type}
}
\value{
Result that with the same type of input
}
\description{
Finding the classes of each variable
}
