% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Scaling the numeric columns of the input x}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble.}

\item{center}{A boolean value that records centering}

\item{scale}{A boolean value that records scaling}
}
\value{
Result with the same type of input
}
\description{
Scaling the numeric columns of the input x
}
