% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a CSV Report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable where the REDCap API token is stored.}

\item{redcapUrl}{The URL of the REDCap API endpoint.}

\item{redcapReportId}{The ID of the report to be downloaded.}
}
\value{
A dataframe containing the data from the specified REDCap report.
}
\description{
This function downloads a specific report from a REDCap project in CSV format
using an API token and report ID. The API token should be stored in an environment
variable whose name is passed to the function. The function returns the report
as a dataframe.
}
\examples{
# Example usage:
# downloadRedcapReport("MY_REDCAP_TOKEN", "https://redcap.example.com/api/", "12345")

}
