% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculating f1 score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A vector that records prediction results}

\item{truth}{A vector that records ground truth label}
}
\value{
A number type result
}
\description{
Calculating f1 score
}
