% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Verify the maximum}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{A boolean value that records if na is allowed}
}
\value{
A logical vector
}
\description{
Verify the maximum
}
\examples{
is_max(c(0,1,1,0,3))
}
