% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{Verify the minimum}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{A boolean value that records if na is allowed}
}
\value{
A logical vector
}
\description{
Verify the minimum
}
\examples{
is_min(c(0,1,1,0,3))
}
