% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Calculate the log-likelihood of x under the chi-square distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{df}{A number that records degrees of freedom.}
}
\value{
A number type result
}
\description{
Calculate the log-likelihood of x under the chi-square distribution
}
