% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Calculating the log-likelihood of x under the f distribution}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{df1}{A number that records degrees of freedom.}

\item{df2}{A number that records degrees of freedom.}
}
\value{
A number type result
}
\description{
Calculating the log-likelihood of x under the f distribution
}
