% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Calculating the log-likelihood of x under the normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{mean}{A number that records mean of x.}

\item{sd}{A number that records standard deviation of x.}
}
\value{
A number type result
}
\description{
Calculating the log-likelihood of x under the normal distribution
}
