% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculating the log-likelihood of x under the t distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{df}{A number that records degrees of freedom.}
}
\value{
A number type result
}
\description{
Calculating the log-likelihood of x under the t distribution
}
