% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculating the minimum n for a t-test with two sample}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{A number type result, Cohen’s d.}

\item{power}{A number type result, whose defaults is 0.8}
}
\value{
A number type result
}
\description{
Calculating the minimum n for a t-test with two sample
}
