% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for One-sample or Two-sample t-tests}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector of data for the first sample.}

\item{x2}{Optional numeric vector of data for the second sample. If not provided,
the function will perform a one-sample t-test analysis assuming the mean
of the population is 0.}
}
\value{
The minimum sample size needed to achieve the desired statistical power and
significance level for the given effect size and test type.
}
\description{
This function calculates the minimum sample size required to achieve a power of 0.80
and a significance level of 0.05 for one-sample or two-sample t-tests, depending on
whether one or two vectors of data are provided.
}
\examples{
# For one-sample t-test
sample1 <- rnorm(50, mean = 5, sd = 1.5)
minimumN(sample1)

# For two-sample t-test
sample2 <- rnorm(50, mean = 5.5, sd = 1.5)
minimumN(sample1, sample2)

}
