% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Approximation of Data}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame. The input data which is to be approximated.}

\item{npc}{The number of principal components to use for the approximation. This number cannot exceed
the number of columns in \code{x}. If it does, the function will stop with an error message.}
}
\value{
A numeric matrix approximating the original data using the specified number of principal components.
The returned data matrix has the same dimensions as the input matrix \code{x} and is scaled back to the
original scale and centering of \code{x}.
}
\description{
This function returns an approximation of the data \code{x} based on a specified number of principal
components (\code{npc}). The approximated data is rescaled and recentered to match the original data dimensions.
The function uses PCA to reduce the dimensionality of the data, retaining only the specified number
of principal components to reconstruct an approximation of the original data.
}
\examples{
data <- matrix(rnorm(100), ncol = 5)
approx_data <- pcApprox(data, 2)
print(approx_data)

}
