% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{Calculating R-squared statistics}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A vector that records prediction results}

\item{truth}{A vector that records ground truth label}
}
\value{
Result of the calculation
}
\description{
Calculating R-squared statistics
}
