% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Repeating the input data based on the number of rows and columns}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or a dataframe.}

\item{M}{A integer that records the number of repeating rows}

\item{N}{A integer that records the number of repeating columns}
}
\value{
Result that with the same type of input
}
\description{
Repeating the input data based on the number of rows and columns
}
