% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{specificity}
\alias{specificity}
\title{Calculating the specificity value}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{A vector that records prediction results}

\item{truth}{A vector that records ground truth label}
}
\value{
A number type result
}
\description{
Calculating the specificity value
}
