% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names}
\usage{
standardizeNames(data, case = "small_camel")
}
\arguments{
\item{data}{A data frame whose variable names are to be standardized.}

\item{case}{A character string specifying the desired case format for variable names.
Supported cases include "small_camel", "big_camel", "snake", etc.}
}
\value{
A data frame with standardized variable names.
}
\description{
This function standardizes the names of variables in a data frame \code{data} to a specified case.
It first converts all variable names to snake case using \code{janitor::clean_names} and then
transforms them to the specified case using \code{snakecase::to_any_case}.
}
\examples{
data <- data.frame(X1 = 1:10, variableTwo = rnorm(10))
standardized_data <- standardizeNames(data)
}
