% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plot a Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numeric vector indicating the event status}

\item{time}{Numeric vector with the time to event or censoring}
}
\value{
A ggplot object representing the survival curve.
}
\description{
This function takes a numerical vector \code{status} and a numerical vector \code{time}, calculates the survival
curve, and plots it using ggplot2.
}
\examples{
status <- c(1, 1, 0, 1, 0, 1)
time <- c(5, 10, 15, 20, 25, 30)
survCurv(status, time)
}
