% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a Scaled Vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled.}
}
\value{
A numeric vector with the scaling and centering reversed.
}
\description{
This function reverses the centering and scaling transformations applied to a vector \code{x}.
}
\examples{
x <- c(1, 2, 3, 4, 5)
scaled_x <- scale(x)
unscale(scaled_x)
}
